requirements.txt
PyQt6==6.10.0
Pillow==12.0.0
